handle exceptions Category-Ingredients

+++ src/main/java/com/pinsoft/internship/recipeshare/service/CategoryService.java
@@ -1,6 +1,7 @@
 package com.pinsoft.internship.recipeshare.service;
 
 import com.pinsoft.internship.recipeshare.entity.Category;
+import com.pinsoft.internship.recipeshare.exceptions.ApiRequestException;
 import com.pinsoft.internship.recipeshare.repository.CategoryRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -12,14 +13,29 @@ import java.util.Optional;
 public class CategoryService {
     @Autowired
     CategoryRepository categoryRepository;
-    public Category add(Category category) {return categoryRepository.save(category);}
+    public Category add(Category category) {
+        if(categoryRepository.findById(category.getId()).isPresent()){
+            throw new ApiRequestException("The given id is already existed!");
+        }else{
+            return categoryRepository.save(category);
+        }
+    }
     public void delete(Long id) {
-        categoryRepository.deleteById(id);
+        if(categoryRepository.findById(id).isEmpty()){
+            throw new ApiRequestException("Category id is not found!");
+        }else{
+            categoryRepository.deleteById(id);
+        }
     }
     public List<Category> getAll(){
         return categoryRepository.findAll();
     }
     public Optional<Category> getById(Long id){
-        return categoryRepository.findById(id);
+        if(categoryRepository.findById(id).isEmpty()){
+            throw new ApiRequestException("The given id is not existed");
+        }else{
+            return categoryRepository.findById(id);
+        }
+
     }
 }
